Modified add_broadcast to notify_add to fit with the new protocol, changed class name accordingly
