+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/di/DataRegistryModule.java
@@ -2,7 +2,6 @@ package com.github.dedis.popstellar.di;
 
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.ACCEPT;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.ADD;
-import static com.github.dedis.popstellar.model.network.method.message.data.Action.ADD_BROADCAST;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.CAST_VOTE;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.CLOSE;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.CREATE;
@@ -11,6 +10,7 @@ import static com.github.dedis.popstellar.model.network.method.message.data.Acti
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.END;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.FAILURE;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.LEARN;
+import static com.github.dedis.popstellar.model.network.method.message.data.Action.NOTIFY_ADD;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.OPEN;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.PREPARE;
 import static com.github.dedis.popstellar.model.network.method.message.data.Action.PROMISE;
@@ -52,7 +52,7 @@ import com.github.dedis.popstellar.model.network.method.message.data.rollcall.Cl
 import com.github.dedis.popstellar.model.network.method.message.data.rollcall.CreateRollCall;
 import com.github.dedis.popstellar.model.network.method.message.data.rollcall.OpenRollCall;
 import com.github.dedis.popstellar.model.network.method.message.data.socialmedia.AddChirp;
-import com.github.dedis.popstellar.model.network.method.message.data.socialmedia.AddChirpBroadcast;
+import com.github.dedis.popstellar.model.network.method.message.data.socialmedia.NotifyAddChirp;
 import com.github.dedis.popstellar.utility.handler.data.ChirpHandler;
 import com.github.dedis.popstellar.utility.handler.data.ConsensusHandler;
 import com.github.dedis.popstellar.utility.handler.data.ElectionHandler;
@@ -123,7 +123,7 @@ public abstract class DataRegistryModule {
     // Social Media
     builder
         .add(CHIRP, ADD, AddChirp.class, ChirpHandler::handleChirpAdd)
-        .add(CHIRP, ADD_BROADCAST, AddChirpBroadcast.class, null);
+        .add(CHIRP, NOTIFY_ADD, NotifyAddChirp.class, null);
 
     return builder.build();
   }
