+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/model/network/method/message/data/socialmedia/NotifyAddChirp.java
@@ -6,7 +6,7 @@ import com.github.dedis.popstellar.model.network.method.message.data.Objects;
 import com.google.gson.annotations.SerializedName;
 
 /** Data sent to broadcast AddChirp to the general channel */
-public class AddChirpBroadcast extends Data {
+public class NotifyAddChirp extends Data {
 
   @SerializedName("post_id")
   private final String postId;
@@ -19,7 +19,7 @@ public class AddChirpBroadcast extends Data {
    * @param channel channel where the post is located
    * @param timestamp UNIX timestamp in UTC
    */
-  public AddChirpBroadcast(String postId, String channel, long timestamp) {
+  public NotifyAddChirp(String postId, String channel, long timestamp) {
     this.postId = postId;
     this.channel = channel;
     this.timestamp = timestamp;
@@ -32,7 +32,7 @@ public class AddChirpBroadcast extends Data {
 
   @Override
   public String getAction() {
-    return Action.ADD_BROADCAST.getAction();
+    return Action.NOTIFY_ADD.getAction();
   }
 
   public String getPostId() {
@@ -55,7 +55,7 @@ public class AddChirpBroadcast extends Data {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    AddChirpBroadcast that = (AddChirpBroadcast) o;
+    NotifyAddChirp that = (NotifyAddChirp) o;
     return java.util.Objects.equals(getPostId(), that.getPostId())
         && java.util.Objects.equals(getChannel(), that.getChannel())
         && java.util.Objects.equals(getTimestamp(), that.getTimestamp());
